+++ src/Symfony/Component/Validator/Constraints/Regex.php
@@ -12,6 +12,7 @@
 namespace Symfony\Component\Validator\Constraints;
 
 use Symfony\Component\Validator\Constraint;
+use Symfony\Component\Validator\Exception\InvalidArgumentException;
 
 /**
  * @Annotation
@@ -31,6 +32,16 @@ class Regex extends Constraint
     public $pattern;
     public $htmlPattern;
     public $match = true;
+    public $normalizer;
+
+    public function __construct($options = null)
+    {
+        parent::__construct($options);
+
+        if (null !== $this->normalizer && !\is_callable($this->normalizer)) {
+            throw new InvalidArgumentException(sprintf('The "normalizer" option must be a valid callable ("%s" given).', \is_object($this->normalizer) ? \get_class($this->normalizer) : \gettype($this->normalizer)));
+        }
+    }
 
     /**
      * {@inheritdoc}