+++ src/Symfony/Component/Validator/CHANGELOG.md
@@ -8,7 +8,8 @@ CHANGELOG
  * added UATP cards support to `CardSchemeValidator`
  * added option `allowNull` to NotBlank constraint
  * added `Json` constraint
- 
+ * added a new `normalizer` option to the string constraints and to the `NotBlank` constraint
+
 4.2.0
 -----